Expand reference table support in subquery pushdown
With this commit, we relax the restrictions put on the reference
tables with subquery pushdown.

Previously, we always expected that the non-reference tables are
equi joined with reference tables on the partition key of the
non-reference table.

With this commit, we allow any column of non-reference tables
joined using non-equi joins as well.